Fix issue where listInOrder function displays only first 10 ids and breaks.
I removed the logic for proceeding down the list and incrementing the id variables from the function definition and placed that block in the main file, after the primary function call. The xId and yId variables are initialized before the function call, and independently of the function definition, although they are defined as function parameters. If user chooses to keep going, the main block increments the id variables and the recalls the function with the new values.